+++ glean/src/core/metrics/types/boolean.ts
@@ -49,7 +49,7 @@ class BooleanMetricType extends MetricType {
   }
 
   /**
-   * **Test-only API**
+   * Test-only API**
    *
    * Gets the currently stored value as a boolean.
    *
@@ -59,7 +59,6 @@ class BooleanMetricType extends MetricType {
    *
    * @param ping the ping from which we want to retrieve this metrics value from.
    *        Defaults to the first value in `sendInPings`.
-   *
    * @returns The value found in storage or `undefined` if nothing was found.
    */
   async testGetValue(ping: string = this.sendInPings[0]): Promise<boolean | undefined> {