+++ glean/src/core/metrics/types/quantity.ts
@@ -4,7 +4,7 @@
 
 import type { CommonMetricData } from "../index.js";
 import { MetricType } from "../index.js";
-import { isNumber } from "../../utils.js";
+import { isInteger } from "../../utils.js";
 import { Context } from "../../context.js";
 import { Metric } from "../metric.js";
 import { ErrorType } from "../../error/error_type.js";
@@ -15,7 +15,7 @@ export class QuantityMetric extends Metric<number, number> {
   }
 
   validate(v: unknown): v is number {
-    if (!isNumber(v)) {
+    if (!isInteger(v)) {
       return false;
     }
 
@@ -87,7 +87,7 @@ class QuantityMetricType extends MetricType {
   }
 
   /**
-   * **Test-only API.**
+   * Test-only API.**
    *
    * Gets the currently stored value as a number.
    *
@@ -97,7 +97,6 @@ class QuantityMetricType extends MetricType {
    *
    * @param ping the ping from which we want to retrieve this metrics value from.
    *        Defaults to the first value in `sendInPings`.
-   *
    * @returns The value found in storage or `undefined` if nothing was found.
    */
   async testGetValue(ping: string = this.sendInPings[0]): Promise<number | undefined> {