+++ glean/src/platform/qt/index.ts
@@ -4,11 +4,10 @@
 
 // Qt does not have its implementations yet, we use pieces of
 // the `TestPlatform` so that the sample will still work.
-import TestPlatform from "../test";
 import uploader from "./uploader.js";
-
-import info from "./platform_info.js";
 import type Platform from "../index.js";
+import TestPlatform from "../test/index.js";
+import info from "./platform_info.js";
 
 const QtPlatform: Platform = {
   Storage: TestPlatform.Storage,